requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Limiter==3.6.0
Werkzeug==3.0.3
python-dotenv